requirements.txt
# requirements.txt
google-generativeai
pylatexenc
python-dotenv